問題リスト関連のUtil関数を括り出した
